Request Limiting

+++ src/components/widgets/buttons/SignOutButton.js
@@ -11,11 +11,19 @@ export default function SignOutButton() {
   const dispatch = useDispatch();
 
   const selectedTarget = useSelector(state => state.selectedTarget);
+  const pendingAPIOperations = useSelector(state => state.pendingAPIOperations);
 
   const signOut = () => {
-    dispatch(actionCreators.resources.clearResources());
-    dispatch(actionCreators.authorization.removeToken(selectedTarget.name));
-    dispatch(actionCreators.targets.clearTarget());
+    if (
+      pendingAPIOperations.indexOf(actionCreators.resources.loadFromTarget.toString()) < 0 &&
+      pendingAPIOperations.indexOf(actionCreators.resources.loadFromTargetSearch.toString()) < 0 &&
+      pendingAPIOperations.indexOf(actionCreators.resources.selectResource.toString()) < 0
+    )
+    {
+      dispatch(actionCreators.resources.clearResources());
+      dispatch(actionCreators.authorization.removeToken(selectedTarget.name));
+      dispatch(actionCreators.targets.clearTarget());
+    }
   };
 
   return (
@@ -25,12 +33,13 @@ export default function SignOutButton() {
       placement="right"
     >
       <ExitToAppIcon
-        css={[
-          buttons.inlineButton,
-          {
-            paddingLeft: 5,
-            verticalAlign: "middle"
-          }]}
+        css={[{paddingLeft: 5, verticalAlign: "middle"},
+          pendingAPIOperations.indexOf(actionCreators.resources.loadFromTarget.toString()) < 0 &&
+          pendingAPIOperations.indexOf(actionCreators.resources.loadFromTargetSearch.toString()) < 0 &&
+          pendingAPIOperations.indexOf(actionCreators.resources.selectResource.toString()) < 0
+            ? [buttons.inlineButton]
+            : [buttons.disabledInlineButton]
+        ]}
         onClick={signOut}
       />
     </Tooltip>
